* Added @showrate to disable/enable the message of rate information when VIP is enabled. (Someone mention about this, I think the message has posibility will be annoying). * Added 'vip_disp_rate' at player.conf to set the message of rate information is displayed or not by default. * Added 'disp_serverbank_msg' at misc.conf to hide "login-serv has been asked to %s the player '%.*s'." message. * Updated map_msg.conf and map_msg_idn.conf

+++ src/login/login.c
@@ -506,7 +506,7 @@ int chrif_sendvipdata(int fd, struct mmo_account acc, char isvip, int mapfd) {
 	WFIFOHEAD(fd,19);
 	WFIFOW(fd,0) = 0x2743;
 	WFIFOL(fd,2) = acc.account_id;
-	WFIFOL(fd,6) = acc.vip_time;
+	WFIFOL(fd,6) = (int)acc.vip_time;
 	WFIFOB(fd,10) = isvip;
 	WFIFOL(fd,11) = acc.group_id; //new group id
 	WFIFOL(fd,15) = mapfd; //link to mapserv
